Fixed update function in crud. Can now create, read, update and delete everything in my database.
Had to create proper type for passing the movieform properties that defines what button should be activated and what function should be run.

Had to fix truthy statements to declare which button should be in operation during which time.

Created clickHandler to choose which function to use during which button was pressed.

Added some CSS stuff for readability.